+++ src/main/java/com/odnolap/tst1/model/dto/AccountDto.java
@@ -4,12 +4,14 @@ import com.odnolap.tst1.model.db.Account;
 import com.odnolap.tst1.model.db.Currency;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class AccountDto {
     private Long id;
     private Long customerId;
     private Currency currency;
-    private Float balance;
+    private BigDecimal balance;
 
     public AccountDto(Account account) {
         if (account != null) {
