+++ src/main/java/com/odnolap/tst1/repository/AccountRepository.java
@@ -2,6 +2,7 @@ package com.odnolap.tst1.repository;
 
 import com.odnolap.tst1.model.db.Account;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public interface AccountRepository {
@@ -9,5 +10,5 @@ public interface AccountRepository {
 
     List<Account> getAllAccounts(int startFrom, int offset);
 
-    boolean moveMoneyBetweenAccounts(Long accountFromId, float amountFrom, Long accountToId, float amountTo);
+    boolean moveMoneyBetweenAccounts(Long accountFromId, BigDecimal amountFrom, Long accountToId, BigDecimal amountTo);
 }
